+++ swanlab/data/run/main.py
@@ -9,7 +9,6 @@ r"""
 """
 import random
 from datetime import datetime
-from enum import Enum
 from typing import Callable, Optional, Dict
 
 from swankit.core import SwanLabSharedSettings
@@ -21,23 +20,12 @@ from swanlab.package import get_package_version
 from . import namer as N
 from .config import SwanLabConfig
 from .exp import SwanLabExp
-from .helper import SwanLabRunOperator, RuntimeInfo
+from .helper import SwanLabRunOperator, RuntimeInfo, SwanLabRunState, MonitorCron, check_log_level
 from .metadata import get_requirements, get_metadata
 from .public import SwanLabPublicConfig
 from ..formater import check_key_format, check_exp_name_format, check_desc_format
 
 MAX_LIST_LENGTH = 108
-
-
-class SwanLabRunState(Enum):
-    """SwanLabRunState is an enumeration class that represents the state of the experiment.
-    We Recommend that you use this enumeration class to represent the state of the experiment.
-    """
-
-    NOT_STARTED = -2
-    SUCCESS = 1
-    CRASHED = -1
-    RUNNING = 0
 
 
 class SwanLabRun:
@@ -101,7 +89,7 @@ class SwanLabRun:
         self.__public = SwanLabPublicConfig(self.__project_name, self.__settings)
         self.__operator.before_run(self.__settings)
         # ---------------------------------- 初始化日志记录器 ----------------------------------
-        swanlog.level = self.__check_log_level(log_level)
+        swanlog.level = check_log_level(log_level)
         # ---------------------------------- 初始化配置 ----------------------------------
         # 如果config是以下几个类别之一，则抛出异常
         if isinstance(run_config, (int, float, str, bool, list, tuple, set)):
@@ -138,6 +126,76 @@ class SwanLabRun:
                 metadata=metadata,
             )
         )
+        # 定时采集系统信息
+        self.monitor_cron = None
+        if self.monitor_funcs is not None and len(self.monitor_funcs) != 0:
+            self.monitor_cron = MonitorCron(self.__get_monitor_func())
+
+    def __get_monitor_func(self):
+        """
+        获取监控函数
+        """
+        if self.monitor_funcs is None or len(self.monitor_funcs) == 0:
+            return None
+
+        def monitor_func():
+            monitor_info_list = [f() for f in self.monitor_funcs]
+            # 剔除其中为None的数据
+            for monitor_info in monitor_info_list:
+                if monitor_info is None:
+                    swanlog.debug("Hardware info is empty. Skip it.")
+                    continue
+                key, name, value = monitor_info['key'], monitor_info['name'], monitor_info['value']
+                v = DataWrapper(key, [Line(value)])
+                self.__exp.add(
+                    data=v,
+                    key=key,
+                    key_name=name,
+                    key_class="SYSTEM",
+                    section_type="SYSTEM",
+                )
+
+        return monitor_func
+
+    def __register_exp(
+        self,
+        experiment_name: str = None,
+        description: str = None,
+        num: int = None,
+    ) -> SwanLabExp:
+        """
+        注册实验，将实验配置写入数据库中，完成实验配置的初始化
+        """
+        if experiment_name:
+            e = check_exp_name_format(experiment_name)
+            if experiment_name != e:
+                swanlog.warning("The experiment name has been truncated automatically.")
+                experiment_name = e
+        if description:
+            d = check_desc_format(description)
+            if description != d:
+                swanlog.warning("The description has been truncated automatically.")
+                description = d
+        experiment_name = N.generate_name(num) if experiment_name is None else experiment_name
+        description = "" if description is None else description
+        colors = N.generate_colors(num)
+        self.__operator.before_init_experiment(self.__run_id, experiment_name, description, num, colors)
+        self.__settings.exp_name = experiment_name
+        self.__settings.exp_colors = colors
+        self.__settings.description = description
+        return SwanLabExp(self.__settings, operator=self.__operator)
+
+    def __stop(self, error: str = None):
+        """
+        停止部分功能，内部清理时调用
+        """
+        if self.monitor_cron is not None:
+            self.monitor_cron.cancel()
+        self.__operator.on_stop(error)
+
+    def __str__(self) -> str:
+        """此类的字符串表示"""
+        return self.__run_id
 
     @property
     def public(self):
@@ -212,7 +270,7 @@ class SwanLabRun:
         _set_run_state(state)
         error = error if state == SwanLabRunState.CRASHED else None
         # 退出回调
-        getattr(run, "_SwanLabRun__operator").on_stop(error)
+        getattr(run, "_SwanLabRun__stop")(error)
         try:
             swanlog.uninstall()
         except RuntimeError:
@@ -311,70 +369,6 @@ class SwanLabRun:
             log_return[metric_info.column_info.key] = metric_info
 
         return log_return
-
-    def __str__(self) -> str:
-        """此类的字符串表示"""
-        return self.__run_id
-
-    def __collect_monitoring_data(self):
-        """
-        采集监控信息，用于监控硬件信息
-        """
-        monitor_info_list = [f() for f in self.monitor_funcs]
-        # 剔除其中为None的数据
-        for monitor_info in monitor_info_list:
-            if monitor_info is None:
-                swanlog.debug("Hardware info is empty. Skip it.")
-                continue
-            key, name, value = monitor_info['key'], monitor_info['name'], monitor_info['value']
-            v = DataWrapper(key, [Line(value)])
-            self.__exp.add(
-                data=v,
-                key=key,
-                key_name=name,
-                key_class="SYSTEM",
-                section_type="SYSTEM",
-            )
-
-    def __register_exp(
-        self,
-        experiment_name: str = None,
-        description: str = None,
-        num: int = None,
-    ) -> SwanLabExp:
-        """
-        注册实验，将实验配置写入数据库中，完成实验配置的初始化
-        """
-        if experiment_name:
-            e = check_exp_name_format(experiment_name)
-            if experiment_name != e:
-                swanlog.warning("The experiment name has been truncated automatically.")
-                experiment_name = e
-        if description:
-            d = check_desc_format(description)
-            if description != d:
-                swanlog.warning("The description has been truncated automatically.")
-                description = d
-        experiment_name = N.generate_name(num) if experiment_name is None else experiment_name
-        description = "" if description is None else description
-        colors = N.generate_colors(num)
-        self.__operator.before_init_experiment(self.__run_id, experiment_name, description, num, colors)
-        self.__settings.exp_name = experiment_name
-        self.__settings.exp_colors = colors
-        self.__settings.description = description
-        return SwanLabExp(self.__settings, operator=self.__operator)
-
-    @staticmethod
-    def __check_log_level(log_level: str) -> str:
-        """检查日志等级是否合法"""
-        valid = ["debug", "info", "warning", "error", "critical"]
-        if log_level is None:
-            return "info"
-        elif log_level.lower() in valid:
-            return log_level.lower()
-        else:
-            swanlog.warning(f"The log level you provided is not valid, it has been set to {log_level}.")
-            return "info"
 
 
 run: Optional["SwanLabRun"] = None
